Entrega 4, autenticación

+++ Backend/src/controllers/audioChannels.js
@@ -7,44 +7,64 @@ const audioChannelController = {
 
     createAudioChannel: (req,res)=>{
         let idGroup = req.params.idGroup;
+        let idUser = req.headers.user;
 
         Group.findById(idGroup)
         .then(grupo =>{
-            //El canal automaticamente se crea con todos los miembros
-            audioChannel.create({arrMembers:grupo.arrUsers})
-            .then(channel => {
-                //Intentamos asignar el _id del canal al arreglo de canales de audio del grupo
-                Group.findByIdAndUpdate(idGroup, {$push:{arrAudioChannels: channel._id}}, { new : true })
-                .then(group => {
-                    res.status(200).type("application/json").json(channel);
+            if(grupo.arrAdmins.includes(idUser)){
+                //El canal automaticamente se crea con todos los miembros
+                audioChannel.create({arrMembers:grupo.arrUsers})
+                .then(channel => {
+                    //Intentamos asignar el _id del canal al arreglo de canales de audio del grupo
+                    Group.findByIdAndUpdate(idGroup, {$push:{arrAudioChannels: channel._id}}, { new : true })
+                    .then(group => {
+                        res.status(200).type("application/json").json(channel);
+                    })
+                    .catch(err => {
+                        //Si no se puedo guardar al grupo el canal, hay que eliminarlo de la base de datos de canales
+                        audioChannel.findByIdAndDelete(channel._id);
+                        res.status(404).send("No se encontró el grupo con el id: "+ idGroup);
+                        
+                    });
                 })
-                .catch(err => {
-                    //Si no se puedo guardar al grupo el canal, hay que eliminarlo de la base de datos de canales
-                    audioChannel.findByIdAndDelete(channel._id);
-                    res.status(404).send("No se encontró el grupo con el id: "+ idGroup);
-                    
-                });
-            })
-            .catch(err =>{
-                res.status(500).send("Error en el servidor " + err);
-            })
+                .catch(err =>{
+                    res.status(500).send("Error en el servidor ");
+                })
+            }
+            else{
+                res.status(403).send('No eres administrador del grupo')
+            }
         })
     },
 
     deleteAudioChannel: (req,res)=>{
         let idGroup = req.params.idGroup;
         let idChannel = req.params.idChannel;
-        Group.findByIdAndUpdate(idGroup,{$pull:{arrAudioChannels: idChannel}})
-        .then(group => {
-            audioChannel.findByIdAndDelete(idChannel)
-            .then(channel => {
-                res.status(200).type("application/json").json(channel);
-            })
-            .catch(err =>{
-                res.status(404).send("No se encontró el canal de audio con el id: "+ idChannel)});
+        let idUser = req.headers.user;
+        Group.findById(idGroup)
+        .then(group=>{
+            if(group.arrAdmins.includes(idUser)){
+                index = group.arrAudioChannels.indexOf(idChannel);
+                group.arrAudioChannels.splice(index,1);
+                group.save()
+                .then(group => {
+                    audioChannel.findByIdAndDelete(idChannel)
+                    .then(channel => {
+                        res.status(200).type("application/json").json(channel);
+                    })
+                    .catch(err =>{
+                        res.status(404).send("No se encontró el canal de audio con el id: "+ idChannel)});
+                })
+                .catch(err => {
+                    res.status(404).send("No se encontró el grupo con el id: "+ idGroup);
+                })
+            }
+            else{
+                res.status(403).send("No eres administrador de este grupo")
+            }
         })
-        .catch(err => {
-            res.status(404).send("No se encontró el grupo con el id: "+ idGroup);
+        .catch(err=>{
+            res.status(404).send("No se encontró el grupo con el id: "+ idGroup)
         })
     },
 
@@ -52,33 +72,44 @@ const audioChannelController = {
         let email = req.body.email;
         let idChannel = req.params.idChannel;
         let idGroup = req.params.idGroup;
-        
+        let idAdmin = req.headers.user;
+
+
+
         //Obtener id con el email
-        User.findOne({email: email})
+        User.findOne({email: email},"-password")
         .then(user=>{
             Group.findById(idGroup)
             .then(group=>{
                 //Validar que forme parte del grupo
-                if(!group.arrUsers.includes(user._id)){
-                    res.status(404).send('El usuario no forma parte del grupo');
+                if(group.arrAdmins.includes(idAdmin)){
+                    if(group.arrUsers.includes(user._id)){
+                        audioChannel.findById(idChannel)
+                        .then(channel=>{
+                            //Validar que no se duplique
+                            if(!channel.arrMembers.includes(user._id)){
+                                channel.arrMembers.push(user._id)
+                                channel.save()
+                                .then(channel=>{
+                                    res.status(200).type("application/json").json(user)
+                                })
+                                .catch(err=>{res.status(400).send("Error al añadir usuario al canal de audio")})
+
+                            }
+                            else{
+                                res.status(400).send('Ya esta agregado el usuario al canal de audio')
+
+                            }
+                        })
+                        .catch(err=>{res.status(404).send('No se encontró al canal de audio con el id '+ idChannel)})
+    
+                    }
+                    else{
+                        res.status(404).send('El usuario no forma parte del grupo');
+                    }
                 }
                 else{
-                    audioChannel.findById(idChannel)
-                    .then(channel=>{
-                        //Validar que no se duplique
-                        if(channel.arrMembers.includes(user._id)){
-                            res.status(400).send('Ya esta agregado el usuario al canal de audio')
-                        }
-                        else{
-                            channel.arrMembers.push(user._id)
-                            channel.save()
-                            .then(channel=>{
-                                res.status(200).type("application/json").json(user)
-                            })
-                            .catch(err=>{res.status(400).send("Error al añadir usuario al canal de audio")})
-                        }
-                    })
-                    .catch(err=>{res.status(404).send('No se encontró al canal de audio con el id '+ idChannel)})
+                    res.status(403).send('No eres administrador del grupo')
                 }
             })
             .catch(err=>{res.status(404).send('No se encontro el grupo con el id '+ idGroup)})
@@ -94,19 +125,32 @@ const audioChannelController = {
         let email = req.body.email;
         let idChannel = req.params.idChannel;
         let idGroup = req.params.idGroup;
-        console.log(email);
-        User.findOne({email: email})
-        .then(user=>{
-            audioChannel.findByIdAndUpdate(idChannel, {$pull:{arrMembers:user._id}}, { new : true })
-            .then(channel => {
-                res.status(200).type("application/json").json(user);
-            })
-            .catch(err => {
-                res.status(404).send("No se encontró el canal de audio con el id: "+ idChannel + " "+ err);
-            });
+        let idAdmin = req.headers.user;
+
+
+        Group.findById(idGroup)
+        .then(group=>{
+            if(group.arrAdmins.includes(idAdmin)){
+                User.findOne({email: email},"-password")
+                .then(user=>{
+                    audioChannel.findByIdAndUpdate(idChannel, {$pull:{arrMembers:user._id}}, { new : true })
+                    .then(channel => {
+                        res.status(200).type("application/json").json(user);
+                    })
+                    .catch(err => {
+                        res.status(404).send("No se encontró el canal de audio con el id: "+ idChannel + " ");
+                    });
+                })
+                .catch(err =>{
+                    res.status(404).send("No se encontró el usuario con el email "+ email + " " );
+                })
+            }
+            else{
+                res.status(403).send("No eres administrador del grupo")
+            }
         })
         .catch(err =>{
-            res.status(404).send("No se encontró el usuario con el email "+ email + " "+err );
+            res.status(404).send("No se encontró el grupo")
         })
     },
 
@@ -119,26 +163,32 @@ const audioChannelController = {
         Group.findById(idGroup)
         .then(group=>{
             //Validar que forme parte del grupo
-            if(!group.arrUsers.includes(idUser)){
-                res.status(404).send('El usuario no forma parte del grupo');
-            }
-            else{
+            if(group.arrUsers.includes(idUser)){
                 audioChannel.findById(idChannel)
                 .then(channel=>{
-                    //Validar que no se duplique
-                    if(channel.arrInCall.includes(idUser)){
-                        res.status(400).send('Ya esta agregado el usuario a la llamada')
+                    if(channel.arrMembers.find(({_id}) => _id == idUser)){
+                        //Validar que no se duplique
+                        if(channel.arrInCall.includes(idUser)){
+                            res.status(400).send('Ya esta agregado el usuario a la llamada')
+                        }
+                        else{
+                            channel.arrInCall.push(idUser)
+                            channel.save()
+                            .then(channel=>{
+                                res.status(200).type("application/json").json(channel)
+                            })
+                            .catch(err=>{res.status(400).send("Error al añadir usuario a la llamada")})
+                        }
                     }
                     else{
-                        channel.arrInCall.push(idUser)
-                        channel.save()
-                        .then(channel=>{
-                            res.status(200).type("application/json").json(channel)
-                        })
-                        .catch(err=>{res.status(400).send("Error al añadir usuario a la llamada")})
+                        res.status(403).send("No formas parte de este canal");
                     }
                 })
                 .catch(err=>{res.status(404).send('No se encontró al canal de audio con el id '+ idChannel)})
+            }
+            else{
+                res.status(404).send('El usuario no forma parte del grupo');
+
             }
         })
         .catch(err=>{res.status(404).send('No se encontro el grupo con el id '+ idGroup)})
@@ -180,24 +230,24 @@ const audioChannelController = {
 
         //verficar si en ese group el usuario es admin
         Group.findById(idGroup)
-            .then(group => {
-                if (group.arrAdmins.includes(idUser)){ //si esta en arreglo de administradores
-                    //se busca el canal a cambiar su nombre 
-                    audioChannel.findByIdAndUpdate(idChannel,{title:newTitle},{ new : true })
-                        .then(updatedChannel => {
-                            res.status(200).type("application/json").json(updatedChannel)
-                        })
-                        .catch(error =>{
-                            res.status(400).send('No pudo modificarse el nombre del canal')
-                        })
-                }
-                else{
-                    res.status(401).send('No tienes permisos para cambiar el nombre')
-                }
-            })
-            .catch(error => {
-                res.status(404).send('No se encontro ese grupo')
-            })
+        .then(group => {
+            if (group.arrAdmins.includes(idUser)){ //si esta en arreglo de administradores
+                //se busca el canal a cambiar su nombre 
+                audioChannel.findByIdAndUpdate(idChannel,{title:newTitle},{ new : true })
+                    .then(updatedChannel => {
+                        res.status(200).type("application/json").json(updatedChannel)
+                    })
+                    .catch(error =>{
+                        res.status(400).send('No pudo modificarse el nombre del canal')
+                    })
+            }
+            else{
+                res.status(403).send('No tienes permisos para cambiar el nombre')
+            }
+        })
+        .catch(error => {
+            res.status(404).send('No se encontro ese grupo')
+        })
     }
 
 }
